feat: add '/api/user/:id' route
