+++ server/src/controllers/postController.ts
@@ -1,21 +1,8 @@
 import { Request, Response } from 'express';
+
 import supabase, { createSupabaseClient } from '../supabaseClient';
 import { generatePublicUrl } from '../utils/publicUrlGenerator';
-
-async function getUsername(userId: string): Promise<string | null> {
-  const { data, error } = await supabase
-    .from('users')
-    .select('username')
-    .eq('id', userId)
-    .single();
-
-  if (error) {
-    console.error('Error fetching username:', error.message);
-    return null;
-  }
-
-  return data?.username || null;
-}
+import { getUsername } from '../utils/userGetter';
 
 export const createPost = async (req: Request, res: Response): Promise<void> => {
   console.log(req.body);
