+++ server/src/utils/userGetter.ts
@@ -1,0 +1,44 @@
+import supabase from '../supabaseClient';
+import { generatePublicUrl } from './publicUrlGenerator';
+
+interface UserProfile {
+  username: string;
+  profilePicture: string | null;
+}
+
+export async function getUser(userId: string): Promise<UserProfile | null> {
+  const { data, error } = await supabase
+    .from('users')
+    .select('username, profile_picture')
+    .eq('id', userId)
+    .single();
+
+  if (error) {
+    console.error('Error fetching username:', error.message);
+    return null;
+  }
+
+  const displayName = data?.username || 'Unknown User';
+
+  let pictureUrl = null;
+  if (data?.profile_picture) {
+    pictureUrl = generatePublicUrl('profile-pictures', data.profile_picture)
+  }
+
+  return { username: displayName, profilePicture: pictureUrl };
+}
+
+export async function getUsername(userId: string): Promise<string | null> {
+  const { data, error } = await supabase
+    .from('users')
+    .select('username')
+    .eq('id', userId)
+    .single();
+
+  if (error) {
+    console.error('Error fetching username:', error.message);
+    return null;
+  }
+
+  return data?.username || null;
+}
